# swagger.yaml
openapi: 3.0.0
info:
  title: Task Manager API
  description: RESTful API for managing user tasks
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /users/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '201':
          description: User registered successfully
        '500':
          description: Failed to register user

  /users/login:
    post:
      summary: Authenticate a user and obtain an access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: User authenticated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Authentication failed
        '500':
          description: Failed to authenticate user

  /tasks:
    post:
      summary: Create a new task
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
              required:
                - title
                - description
      responses:
        '201':
          description: Task created successfully
        '401':
          description: Unauthorized
        '500':
          description: Failed to create task

    get:
      summary: Get all tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized
        '500':
          description: Failed to fetch tasks

  /tasks/{taskId}:
    put:
      summary: Update an existing task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          description: ID of the task to update
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '500':
          description: Failed to update task

    delete:
      summary: Delete an existing task
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          description: ID of the task to delete
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted successfully
        '401':
          description: Unauthorized
        '404':
          description: Task not found
        '500':
          description: Failed to delete task

components:
  schemas:
    Task:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        userId:
          type: string
      required:
        - _id
        - title
        - description
        - userId
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
      required:
        - title
        - description
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT